Drop dependency on GNU libunwind shared library.
We would like to use the name "libunwind" for something other than the
almost-unused GNU libunwind. This is the only user of the GNU libunwind
shared library in the tree, aside from GNU libunwind's own unit tests (which
I will fix separately). However, the code that uses GNU libunwind was put
behind an #ifndef ANDROID in commit 57eb94f1. The only remaining use was
the reference in the header #include. Put that behind the #ifndef as well,
so that we can remove the reference to libunwind.

Bug: 144430859
Change-Id: Ic5049762d9e3fa11beff946927baa771f7152ff5