Fix igt_gpu_tools test: kms_atomic#test_only fails
Root cause: If a pipe has multiple primary plane, the test will always
fail. Because if the pipe's mode is changed, there will be commit for
each of it's primary plane. In this subtest, first primary plane has
some change that we want to in the legacy way. But the second primary
plane is also making a commit, which is not intended.

Fix: If the mode of a pipe is changed, make the commit only for the first primary plane on the pipe.

Bug: b/157692504
Test: run kms_atomic test
Change-Id: I332aec022a5250342f789b450ffe92e97f06e0e0