Fix igt-gpu-tools/kms_vblank test failure
Root cause: Legacy commit on a pipe with muliple primary plane. In the legacy commit for a primary plane, it overrides the current primary plane on the crtc, with the call to drmModeSetCrtc(). So, for a device with multiple primary plane, the plane will always override the first primary planes content.

Fix: For the legacy commit, treat only the first primary plane as a primary plane and all the remaining primary as overlay plane.

For hanging subtest, changed it to fail instead of hang.

Bug: b/254180219
Test: run kms_vblank test
Change-Id: I29ae19619a459540723d737495956f5123e031bc